+++ hcloud/schema.go
@@ -212,7 +212,7 @@ func ServerPrivateNetFromSchema(s schema.ServerPrivateNet) ServerPrivateNet {
 	n := ServerPrivateNet{
 		Network:    &Network{ID: s.Network},
 		IP:         net.ParseIP(s.IP),
-		MacAddress: s.MacAddress,
+		MACAddress: s.MACAddress,
 	}
 	for _, ip := range s.AliasIPs {
 		n.Aliases = append(n.Aliases, net.ParseIP(ip))