+++ hcloud/schema_test.go
@@ -661,8 +661,8 @@ func TestServerPrivateNetFromSchema(t *testing.T) {
 	if privateNet.Aliases[0].String() != "10.0.1.2" {
 		t.Errorf("unexpected alias IP: %v", privateNet.Aliases[0])
 	}
-	if privateNet.MacAddress != "86:00:ff:2a:7d:e1" {
-		t.Errorf("unexpected mac address: %v", privateNet.MacAddress)
+	if privateNet.MACAddress != "86:00:ff:2a:7d:e1" {
+		t.Errorf("unexpected mac address: %v", privateNet.MACAddress)
 	}
 }